REFACTOR: rename cache.py to header_auth_cache.py for clarity
The module contains authorization caching logic specific to the header-based
auth feature (/check-header endpoint). The new name better reflects its
purpose and leaves room for other caching modules in the future.
